refactor(Settings): Moves roles and users pages into pages folder.
